+++ Core/Util/Configs/ConfigEnums.cs
@@ -3,8 +3,8 @@ using Helion.Render.Common.Textures;
 using Helion.Resources.Definitions;
 using Helion.Util.Config.Components;
 using Helion.Util.Configs.Components;
+using Helion.Util.RandomGenerators;
 using Helion.World;
-using Helion.World.Entities.Inventories;
 using Helion.World.Entities.Players;
 using Helion.World.StatusBar;
 using OpenTK.Windowing.Common;
@@ -42,7 +42,8 @@ namespace Helion.Util.Configs
             { typeof(RenderColorMode), Enum.GetValues<RenderColorMode>() },
             { typeof(BlitFilter), Enum.GetValues<BlitFilter>() },
             { typeof(GyroTurnAxis), Enum.GetValues<GyroTurnAxis>() },
-            { typeof(ConfigWeaponSlots), Enum.GetValues<ConfigWeaponSlots>() }
+            { typeof(ConfigWeaponSlots), Enum.GetValues<ConfigWeaponSlots>() },
+            { typeof(RngMethod), Enum.GetValues<RngMethod>() }
         };
 
         public static Dictionary<Type, Dictionary<Enum, string>> KnownEnumLabels { get; } = new Dictionary<Type, Dictionary<Enum, string>>()
@@ -64,7 +65,8 @@ namespace Helion.Util.Configs
             { typeof(RenderColorMode), GetDescriptions<RenderColorMode>() },
             { typeof(BlitFilter), GetDescriptions<BlitFilter>() },
             { typeof(GyroTurnAxis), GetDescriptions<GyroTurnAxis>() },
-            { typeof(ConfigWeaponSlots), GetDescriptions<ConfigWeaponSlots>()}
+            { typeof(ConfigWeaponSlots), GetDescriptions<ConfigWeaponSlots>()},
+            { typeof(RngMethod), GetDescriptions<RngMethod>()}
         };
 
         private static Dictionary<Enum, string> GetDescriptions<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicFields)] T>() where T : struct, Enum