+++ Core/Util/SerializationContexts/WorldModelSerializationContext.cs
@@ -1,6 +1,7 @@
 ﻿namespace Helion.Util.SerializationContexts
 {
     using Helion.Models;
+    using Helion.Util.RandomGenerators;
     using System.Text.Json.Serialization;
 
     [JsonSourceGenerationOptions(
@@ -20,6 +21,7 @@
     [JsonSerializable(typeof(SectorMoveSpecialModel), TypeInfoPropertyName = "SectorMoveSpecialModel")]
     [JsonSerializable(typeof(StairSpecialModel), TypeInfoPropertyName = "StairSpecialModel")]
     [JsonSerializable(typeof(SwitchChangeSpecialModel), TypeInfoPropertyName = "SwitchChangeSpecialModel")]
+    [JsonSerializable(typeof(RngMethod), TypeInfoPropertyName = "RandomMethod")]
     public partial class WorldModelSerializationContext : JsonSerializerContext
     {
     }